requirements.txt
altair==5.1.2
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
DateTime==5.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
eventemitter==0.2.0
exceptiongroup==1.1.3
exchange-calendars==4.5
executing==2.0.1
fastjsonschema==2.18.1
fonttools==4.44.1
fqdn==1.5.1
frozendict==2.3.8
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
h2==3.2.0
hpack==3.0.0
html5lib==1.1
httpcore==0.17.2
httpx==0.24.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.23.4
overrides==7.4.0
packaging==23.2
pandas==2.0.2
pandas-market-calendars==4.3.1
pandocfilters==1.5.0
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
Pillow==10.1.0
pipreq==0.4
platformdirs==4.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pycparser==2.21
pydeck==0.8.1b0
pyee==9.0.4
Pygments==2.16.1
pyhumps==3.0.2
pyluach==2.2.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-configuration==0.9.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
refinitiv-data==1.5.0
refinitiv-dataplatform==1.0.0a21
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.12.0
scipy==1.9.3
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.28.1
tenacity==8.0.1
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.18
validators==0.22.0
watchdog==2.1.9
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.9
yfinance==0.2.31
zipp==3.17.0
zope.interface==6.1
